+++ src/Cirreum.RemoteConnections.WebSockets/RemoteConnectionEnvelope.cs
@@ -1,4 +1,4 @@
-namespace Cirreum.RemoteServices;
+﻿namespace Cirreum.RemoteServices.Connections;
 
 using System.Text.Json;
 using System.Text.Json.Serialization;