+++ src/Cirreum.RemoteConnections.WebSockets/WebSocketRemoteConnection.cs
@@ -1,4 +1,4 @@
-namespace Cirreum.RemoteServices;
+﻿namespace Cirreum.RemoteServices.Connections;
 
 using System.Collections.Concurrent;
 using System.Net.WebSockets;